+++ js/constants.js
@@ -1,0 +1,58 @@
+export const COMMENTS_NUMBER = 30;
+export const PHOTOS_NUMBER = 25;
+export const MIN_COMMENT_ID = 1;
+export const MIN_PHOTO_ID = 1;
+export const MIN_LIKES_NUMBER = 25;
+export const MAX_LIKES_NUMBER = 200;
+export const MIN_AVATAR_ID = 1;
+export const MAX_AVATAR_ID = 6;
+
+export const COMMENT_MESSAGES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+];
+
+export const USER_NAMES = [
+  'Ульяна',
+  'Мирослава',
+  'Полина',
+  'Владимир',
+  'Маргарита',
+  'Иван',
+  'Виктория',
+  'Ясмина',
+  'Аделина',
+  'Милана'
+];
+
+export const PHOTO_DESCRIPTIONS = [
+  'Красивый песчаный пляж с шезлонгами, на заднем плане пятизвездочный отель в окружении леса',
+  'Деревянный знак "Идите на пляж"',
+  'Каменистый пляж с водоемом и деревьями. Он демонстрирует природный ландшафт, где скалы встречаются с водой, на фоне деревьев',
+  'Девушка на пляже в купальнике с фотоаппаратом. Она стоит на песке на фоне океана',
+  '2 миски супа с рисом. Необычная подача - рис в виде человечков, которые лежал в миске, как в ванне',
+  'Черная матовая Ламборгини с открытой дверцей',
+  'Деревянная тарелка с разрезанной пополам ягодой клубники и вилкой',
+  'На столе стоят 2 стеклянных кружки с клюквенным соком, рядом лежат веточки клюквы',
+  'Девушка на пляже пытается поймать самолет, пролетающий над ней',
+  'Выдвинутая обувница на колесиках с 3 парами обуви',
+  'Песочный пляж с голубым небом и забором, который огораживает растительность',
+  'Белая ауди на асфальте',
+  'Нарезка из сельдерея, рыбы, моркови и зеленого перца',
+  'Рыжий котик лежит на рисе в виде сашими',
+  'Ноги в тапочках, которые выглядят как сапоги астронавта, лежат на ручке белого дивана. Рядом стоит тумбочка с книгой, орхидеей и включеной белой лампой',
+  'Вид из окна самолета на котором виден другой самолет, внизу облака, а под облаками горы',
+  'Оркестр в темной одежде стоит на сцене и что-то исполняет',
+  'Красная Импала стоит в кирпичном гараже',
+  'Тапочки, в которые встроены фонарики, светят в темноте и освещают тумбочку',
+  'Курорта в сумерки, на переднем плане пальмы и подсвеченые домики',
+  'Тарелка с салатом из курицы и долькой лайма, рядом лежит вилка',
+  'Закат над морем',
+  'Краб на бревне',
+  'Вид с концерта, где много людей',
+  'Белый джип заехал в водоем, рядом гиппопотам, который разинул пасть'
+];
